+++ Src/Common/W3C/W3CConstants.cs
@@ -1,19 +1,17 @@
 ﻿#if DEPENDENCY_COLLECTOR
     namespace Microsoft.ApplicationInsights.W3C
 #else
-namespace Microsoft.ApplicationInsights.W3C.Internal
+    namespace Microsoft.ApplicationInsights.W3C.Internal
 #endif
 {
-    using System;
     using System.ComponentModel;
 
     /// <summary>
     /// W3C constants.
     /// </summary>
-    [Obsolete("Not ready for public consumption.")]
     [EditorBrowsable(EditorBrowsableState.Never)]
 #if DEPENDENCY_COLLECTOR
-    public 
+    public
 #else
     internal
 #endif