+++ Src/DependencyCollector/Shared/Implementation/ClientServerDependencyTracker.cs
@@ -4,9 +4,8 @@
     using System.Data.SqlClient;
     using System.Diagnostics;
     using System.Net;
-    using Microsoft.ApplicationInsights.Common;
     using Microsoft.ApplicationInsights.DataContracts;
-    using Microsoft.ApplicationInsights.W3C;
+    using Microsoft.ApplicationInsights.Extensibility.W3C;
 
     internal static class ClientServerDependencyTracker
     {
@@ -72,7 +71,7 @@
                 // with W3C support on .NET https://github.com/dotnet/corefx/issues/30331 (TODO)
                 if (currentActivity == null)
                 {
-                    activity.SetParentId(StringUtilities.GenerateTraceId());
+                    activity.SetParentId(W3CUtilities.GenerateTraceId());
                 }
 
                 // end of workaround
@@ -82,9 +81,7 @@
 
             if (IsW3CEnabled)
             {
-#pragma warning disable 612, 618
-                W3COperationCorrelationTelemetryInitializer.UpdateTelemetry(telemetry, activity, true);
-#pragma warning restore 612, 618
+                activity.UpdateTelemetry(telemetry, true);
             }
             else
             {