+++ Src/DependencyCollector/Shared/Implementation/HttpProcessing.cs
@@ -11,7 +11,7 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
     using Microsoft.ApplicationInsights.DataContracts;
     using Microsoft.ApplicationInsights.Extensibility;
     using Microsoft.ApplicationInsights.Extensibility.Implementation;
-    using Microsoft.ApplicationInsights.W3C;
+    using Microsoft.ApplicationInsights.Extensibility.W3C;
 
     /// <summary>
     /// Concrete class with all processing logic to generate RDD data from the callbacks
@@ -210,17 +210,16 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
                         }
                     }
 
-#pragma warning disable 612, 618
                     if (this.injectW3CHeaders && currentActivity != null)
                     {
                         string traceParent = currentActivity.GetTraceparent();
-                        if (traceParent != null && webRequest.Headers[W3CConstants.TraceParentHeader] == null)
+                        if (traceParent != null && webRequest.Headers[W3C.W3CConstants.TraceParentHeader] == null)
                         {
-                            webRequest.Headers.Add(W3CConstants.TraceParentHeader, traceParent);
+                            webRequest.Headers.Add(W3C.W3CConstants.TraceParentHeader, traceParent);
                         }
 
                         string traceState = currentActivity.GetTracestate();
-                        if (webRequest.Headers[W3CConstants.TraceStateHeader] == null)
+                        if (webRequest.Headers[W3C.W3CConstants.TraceStateHeader] == null)
                         {
                             if (applicationId != null)
                             {
@@ -238,11 +237,10 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
 
                             if (traceState != null)
                             {
-                                webRequest.Headers.Add(W3CConstants.TraceStateHeader, traceState);
+                                webRequest.Headers.Add(W3C.W3CConstants.TraceStateHeader, traceState);
                             }
                         }
                     }
-#pragma warning restore 612, 618
                 }
             }
             catch (Exception exception)