+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 2.9.0
 - [Fix: remove unused reference to Microsoft.AspNet.TelemetryCorrelation package from DependencyCollector](https://github.com/Microsoft/ApplicationInsights-dotnet-server/pull/1136)
+- [Move W3C support from DependencyCollector package to base SDK, deprecate W3C support in DependencyCollector](https://github.com/Microsoft/ApplicationInsights-dotnet-server/pull/1138)
 
 ## Version 2.9.0-beta3
 - Update Base SDK to version 2.9.0-beta3